add types and reorganize arcpy functions, try except arcpy import
